добавлен блок с комментариями
